pyproject.toml
[tool.poetry]
name = "ContextRAG"
version = "0.0.1"
description = ""
authors = ["Sean Brar <hello@seanbrar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "8.1.7"
joblib = "1.3.2"
nltk = "3.8.1"
regex = "2023.10.3"
tqdm = "4.66.1"
gensim = "^4.3.2"
beautifulsoup4 = "^4.12.2"
html2text = "^2020.1.16"
pytest = "^7.4.4"
scikit-learn = "^1.3.2"
python-dotenv = "^1.0.0"
openai = "^1.7.0"
tiktoken = ">=0.5.2,<0.7.0"
chromadb = "^0.4.24"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
